Create files for viewing code
Previously was referencing a link on github to view the site.  Because our code for the project has to be private (school rules) we have to access the code locally.  We can do it this way or import a .js library. Depending on the time it takes for proper implementation, I will try to do both.